+++ src/Tabs/index.tsx
@@ -1,0 +1,6 @@
+import { Tabs } from 'react-tabs';
+import { Tab } from './Tab';
+import { TabList } from './TabList';
+import { TabPanel } from './TabPanel';
+
+export { Tabs, TabList, Tab, TabPanel };
